Organize components and Sass files into appropriate folders
