+++ src/Components/MovieCard/MovieCard.scss
