+++ src/Components/MovieSection/MovieSection.scss
