+++ src/Components/Nav/Nav.scss
@@ -1,4 +1,4 @@
-@use "mixins.scss" as mixin;
+@use "src/mixins.scss" as mixin;
 
 nav {
     width: 100%;
